+++ examples/rest-spring-3-example/src/main/java/com/accenture/examples/rest/HelloPoJoServiceMesh.java
@@ -46,7 +46,7 @@ public class HelloPoJoServiceMesh {
                     SamplePoJo pojo = response.getBody(SamplePoJo.class);
                     callback.success(pojo);
                 } else {
-                    callback.error(new AppException(response.getStatus(), response.getError()));
+                    callback.error(new AppException(response.getStatus(), String.valueOf(response.getError())));
                 }
             } catch (IOException | ExecutionException | InterruptedException e) {
                 callback.error(e);